requirements.txt
pillow>=9.0.0
requests>=2.28.0
python-dotenv>=0.19.0
google-cloud-vision>=3.0.0
azure-cognitiveservices-vision-computervision>=0.9.0
PyYAML>=6.0
img2pdf>=0.4.4